Rework the tute to showcase metadata
And also how to edit each row (using metadata as an example), using the new tskit 0.4 accessors.

I also changed round the order so that we illustrate accessing columns as a big numpy array (which is a common thing to want to do) before accessing / writing to rows, as the logical flow worked better like that. The example for major edits showcases how to map mutations parsimoniously, rather than removing singletons, which is now done more easily using `delete_sites` (example also included)